# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminRegionOut": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "country_code": {
            "title": "Country Code",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/AdminZoneOut"
            },
            "title": "Zones",
            "type": "array"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "country_code",
          "description",
          "active",
          "zones"
        ],
        "title": "AdminRegionOut",
        "type": "object"
      },
...
      "AdminZoneOut": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "clusters": {
            "items": {
              "type": "string"
            },
            "title": "Clusters",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "region_id": {
            "format": "uuid",
            "title": "Region Id",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "usable": {
            "title": "Usable",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "region_id",
          "active",
          "clusters",
          "usable"
        ],
        "title": "AdminZoneOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegionCreate": {
        "additionalProperties": false,
        "properties": {
          "country_code": {
            "pattern": "^[A-Za-z]{2}$",
            "title": "Country Code",
            "type": "string"
          },
          "description": {
            "default": "",
            "maxLength": 500,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "pattern": "^[a-z0-9]([a-z0-9-]{0,30}[a-z0-9])?$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "country_code"
        ],
        "title": "RegionCreate",
        "type": "object"
      },
...
      "RegionUpdate": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "country_code": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z]{2}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "RegionUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "ZoneCreate": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "default": "",
            "maxLength": 500,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "pattern": "^[a-z0-9]([a-z0-9-]{0,30}[a-z0-9])?$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name"
        ],
        "title": "ZoneCreate",
        "type": "object"
      },
...
      "ZoneUpdate": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "ZoneUpdate",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Tenant-scoped routes (projects, instances, jobs, role bindings, dashboard) require the header `X-Tenant-Id: <tenant uuid>`; it is validated against the caller's memberships. It is omitted from each operation so generated clients can inject it once.",
    "title": "Cloud Manager API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/regions": {
      "get": {
        "operationId": "admin_list_regions_api_v1_admin_regions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminRegionOut"
                  },
                  "title": "Response Admin List Regions Api V1 Admin Regions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Admin List Regions",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "create_region_api_v1_admin_regions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRegionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Region",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/regions/{region_id}": {
      "patch": {
        "operationId": "update_region_api_v1_admin_regions__region_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "region_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Region Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRegionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Region",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/regions/{region_id}/zones": {
      "post": {
        "operationId": "create_zone_api_v1_admin_regions__region_id__zones_post",
        "parameters": [
          {
            "in": "path",
            "name": "region_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Region Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZoneCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRegionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Zone",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/admin/zones/{zone_id}": {
      "patch": {
        "operationId": "update_zone_api_v1_admin_zones__zone_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "zone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Zone Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZoneUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRegionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Zone",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}